Enock job seeker

+++ db/migrate/20230729203627_job_seeker.rb
@@ -3,7 +3,7 @@ class JobSeeker < ActiveRecord::Migration[7.0]
     create_table :job_seekers do |t|
       t.string :profile_image
       t.string :name
-      t.integer :contact
+      t.bigint :contact
       t.string :resume_attachment
       t.belongs_to :user, null: false, foreign_key: true
